Copying Stack tags to the Datadog monitor (#226)
Cloudformation supports tags at a stack level. They are normally passed on to all Cloudformation resources (some resources don't support that yet...). Datadog monitors got tags, too. With this change, the Cloudformation Stack tags are copied to the Datadog monitor. Additionally, the AWS autogenerated tags like stack name and logical id are also set. If a tag is already set explicitly at a monitor, the monitor value is being used.
Since Datadog uses a : to separate key and value, all : are replaced with _.

In order to maintain backwards compatibility, this feature is switched off by default. It can be enabled account-wide via the type configuration, the same way Datadog credentials are being set.